pull up a few small rules
